api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Hilmi <hilmusyaf2005@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
firebase-admin = "^6.5.0"
pydantic = "^2.8.2"
google-cloud-firestore = "^2.16.1"
boto3 = "^1.34.144"
pillow = "^10.4.0"
google-generativeai = "^0.7.2"
instaloader = "^4.12"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
taskipy = "^1.13.0"
ruff = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
py_version = 312
split_on_trailing_comma = true
force_alphabetical_sort_within_sections = true
multi_line_output = 5
group_by_package = true
line_length = 100
skip_gitignore = true

[tool.taskipy.tasks]
dev = "fastapi dev app/main.py"
format = "ruff check app --fix && ruff format app && isort app"